Upload List ADT File
List.java: List.java is an Abstract Data Type that is composed of integers. The list acts as a bi-directional queue. There are different methods that allow the list of integers (stored in nodes) to be manipulated. There is also a cursor node that highlights a specific node. The cursor has an index which allows for iteration through the list. The cursor can be in the list or can be absent.